- replace actions/selectors from old reducer by the new slices one. Rebase dev
- fix osrdConfContext

- add mapInfraID in map store

- add infraID in props TracksGeographic

- add infraID in props common/Map/Layers/Routes

- add infraID in props common/Map/Layers/OperationalPoints

- add infraID in props common/Map/Layers/Catenaries

- add infraID in props BufferStops, Detectors, LineSearchLayer, NeutralSections, Signals, SpeedLimits, Switches, SNCF_PSL

- add infraID in props common/Map/Layers/GeoJSONs and add editorInfraID in editor slice

- remove eslint warning

- create commonState builder

- change osrdConfContext to osrdContext

- create mapViewer slice

- fix eslint error

- fix error after rebase

- add test mapViewer

- fix review's comments

front: change type snakeCase to PascalCase and extends InfraStateType to StdcmConf and OperationalStudyConf

front: create and use useOsrdConfSelector

front: create and use useInfraID

front: create and use useOsrdConfActions

front: create and use useUpdateInfraID

front: move featureInfoClick from infra reducer to osrdConf reducer

front: rename/harmonize

front: remove stdcmMode from store

front: fix

front: remove unnecessary condition

front: typo

front: fix

front: remove unised cast

front: disable immutableCheck

+++ front/src/common/InfraSelector/InfraSelectorModalBodyStandard.tsx
@@ -1,13 +1,20 @@
 import React, { useContext } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useDispatch } from 'react-redux';
 import nextId from 'react-id-generator';
 import { GoLock } from 'react-icons/go';
+
+import { MODES } from 'applications/operationalStudies/consts';
+
 import InputSNCF from 'common/BootstrapSNCF/InputSNCF';
-import { useDispatch, useSelector } from 'react-redux';
-import { getInfraID } from 'reducers/osrdconf/selectors';
-import { updateInfraID, deleteItinerary } from 'reducers/osrdconf';
-import { useTranslation } from 'react-i18next';
+import type { Infra } from 'common/api/osrdEditoastApi';
 import { ModalContext } from 'common/BootstrapSNCF/ModalSNCF/ModalProvider';
-import { Infra } from 'common/api/osrdEditoastApi';
+import {
+  useInfraID,
+  useOsrdConfActions,
+  useOsrdContext,
+  useUpdateInfraID,
+} from 'common/osrdContext';
 
 type InfraSelectorModalBodyStandardProps = {
   filter: string;
@@ -38,13 +45,16 @@ export default function InfraSelectorModalBodyStandard({
 }: InfraSelectorModalBodyStandardProps) {
   const { t } = useTranslation(['translation', 'infraManagement']);
   const dispatch = useDispatch();
-  const infraID = useSelector(getInfraID);
+  const { mode } = useOsrdContext();
+  const updateInfraID = useUpdateInfraID();
+  const infraID = useInfraID();
   const { closeModal } = useContext(ModalContext);
+  const { deleteItinerary } = useOsrdConfActions();
 
   function setInfraID(id: number) {
     dispatch(updateInfraID(id));
     if (onInfraChange) onInfraChange(id);
-    dispatch(deleteItinerary());
+    if ([MODES.simulation, MODES.stdcm].includes(mode)) dispatch(deleteItinerary());
     if (!onlySelectionMode) {
       closeModal();
     }
